moved waste emissions functions to own class
